タスクごとに同時編集できるように修正
