+++ src/components/organisms/Tasks.tsx
@@ -1,0 +1,15 @@
+import { Task } from "../molecules/Task";
+export const Tasks = (props) => {
+  const { todos, setTodos, onClickCheck, onClickDelete } = props;
+
+  return todos.map((todo) => (
+    <Task
+      key={todo.id}
+      todos={todos}
+      setTodos={setTodos}
+      todo={todo}
+      onClickCheck={onClickCheck}
+      onClickDelete={onClickDelete}
+    />
+  ));
+};
